2. Тесты для идеальный вес [#326977]

+++ src/ru/job4j/calculator/Fit.java
@@ -1,18 +1,18 @@
 package ru.job4j.calculator;
 
 public class Fit {
-    public static double manWeight(short height) {
+    public static double manWeight(double height) {
         double rsl = (height - 100) * 1.15;
         return rsl;
     }
 
-    public static double womanWeight(short height) {
+    public static double womanWeight(double height) {
         double rsl =  (height - 110) * 1.15;
         return rsl;
     }
 
     public static void main(String[] args) {
-        short height = 178;
+        double height = 178;
         double man = Fit.manWeight(height);
         double woman = Fit.womanWeight(height);
         System.out.println(man + " is best weight for man with height 178");
